Move and rename many files
