+++ randomizer/Patching/BossRando.py
@@ -1,5 +1,5 @@
 """Apply Boss Locations."""
-from randomizer.Patcher import ROM
+from randomizer.Patching.Patcher import ROM
 from randomizer.Spoiler import Spoiler
