+++ randomizer/Patching/CosmeticColors.py
@@ -1,5 +1,5 @@
 """Apply cosmetic skins to kongs."""
-from randomizer.Patcher import ROM
+from randomizer.Patching.Patcher import ROM
 from randomizer.Spoiler import Spoiler
