+++ randomizer/Patching/EnemyRando.py
@@ -2,9 +2,9 @@
 import random
 
 import js
-from randomizer.EnemyTypes import Enemies, EnemyMetaData
-from randomizer.MapsAndExits import Maps
-from randomizer.Patcher import ROM
+from randomizer.Lists.EnemyTypes import Enemies, EnemyMetaData
+from randomizer.Lists.MapsAndExits import Maps
+from randomizer.Patching.Patcher import ROM
 from randomizer.Spoiler import Spoiler
