+++ randomizer/ShuffleWarps.py
@@ -4,7 +4,7 @@ import random
 import js
 from randomizer.Enums.Warps import Warps
 from randomizer.Lists.Warps import BananaportVanilla
-from randomizer.MapsAndExits import Maps
+from randomizer.Lists.MapsAndExits import Maps
 
 
 def getShuffleMaps():
